Improved notifications structure
